Factor out `FqVector::entry`
